Resolve potential security vulnerabilities
 - Used `uint64` instead of `int` for Date to comply with G115 (CWE-190)
 - Utilised `filepath.Clean` and suppressed G304 (CWE-22)
 - Configured `ReadHeaderTimeout` in the http.Server to comply with G112 (CWE-400)
 - Reduced directory permission to `750` for complying with G301 (CWE-276)

+++ metasource/models/home/base.go
@@ -71,7 +71,7 @@ type OthrUnit struct {
 	Key    int
 	Author sql.NullString
 	Text   sql.NullString
-	Date   int
+	Date   uint64
 }
 
 type OthrRslt struct {
